+++ CHANGELOG.md
@@ -98,6 +98,7 @@
 # Future features to develop
 - Player vs npc fighting.
 - Player vs player fighting.
+- Refactor create-wisp, create-evil-wisp, create-blessed-wisp, npc, and player as ES6 class files instead of functions.
 - Improve time to interactive by optimizing images format. See Chrome Audit.
 - Fix the blinking of sprites and make their movement smoother.
 - Fix some sprites that are not fighting, not dead but not moving either.